+++ bigquery/src/main/scala_2.13+/org/datatools/bigdatatypes/bigquery/JavaConverters.scala
@@ -1,10 +1,10 @@
 package org.datatools.bigdatatypes.bigquery
 
 import java.lang
-
-import scala.jdk.CollectionConverters.IterableHasAsJava
+import scala.jdk.CollectionConverters.{IterableHasAsJava, IterableHasAsScala}
 
 object JavaConverters {
 
   def toJava[A](value: List[A]): lang.Iterable[A] = value.asJava
+  def toScala[A](value: lang.Iterable[A]): List[A] = value.asScala.toList
 }